Refactor namespaces according to ADR-0002
