+++ explainability-core/src/main/java/org/kie/trustyai/explainability/local/lime/optim/LimeImpactScoreCalculator.java
@@ -25,10 +25,10 @@ import java.util.concurrent.TimeoutException;
 import org.kie.trustyai.explainability.Config;
 import org.kie.trustyai.explainability.local.lime.LimeConfig;
 import org.kie.trustyai.explainability.local.lime.LimeExplainer;
-import org.kie.trustyai.explainability.metrics.ExplainabilityMetrics;
 import org.kie.trustyai.explainability.model.FeatureImportance;
 import org.kie.trustyai.explainability.model.Prediction;
 import org.kie.trustyai.explainability.model.Saliency;
+import org.kie.trustyai.metrics.explainability.ExplainabilityMetrics;
 import org.optaplanner.core.api.score.buildin.simplebigdecimal.SimpleBigDecimalScore;
 import org.optaplanner.core.api.score.calculator.EasyScoreCalculator;
 import org.slf4j.Logger;
