+++ explainability-core/src/main/java/org/kie/trustyai/explainability/local/lime/optim/LimeStabilityScoreCalculator.java
@@ -23,10 +23,10 @@ import java.util.concurrent.TimeoutException;
 
 import org.kie.trustyai.explainability.local.lime.LimeConfig;
 import org.kie.trustyai.explainability.local.lime.LimeExplainer;
-import org.kie.trustyai.explainability.metrics.ExplainabilityMetrics;
 import org.kie.trustyai.explainability.model.Prediction;
 import org.kie.trustyai.explainability.model.PredictionProvider;
 import org.kie.trustyai.explainability.utils.LocalSaliencyStability;
+import org.kie.trustyai.metrics.explainability.ExplainabilityMetrics;
 import org.optaplanner.core.api.score.buildin.simplebigdecimal.SimpleBigDecimalScore;
 import org.optaplanner.core.api.score.calculator.EasyScoreCalculator;
 import org.slf4j.Logger;
