+++ explainability-core/src/main/java/org/kie/trustyai/metrics/accuracy/ConfusionMatrix.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.trustyai.explainability.metrics;
+package org.kie.trustyai.metrics.accuracy;
 
 public class ConfusionMatrix {
