+++ explainability-core/src/main/java/org/kie/trustyai/metrics/fairness/group/GroupAverageOddsDifference.java
@@ -1,4 +1,4 @@
-package org.kie.trustyai.explainability.metrics.fairness.group;
+package org.kie.trustyai.metrics.fairness.group;
 
 import java.util.List;
 import java.util.Map;
@@ -6,19 +6,19 @@ import java.util.concurrent.ExecutionException;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-import org.kie.trustyai.explainability.metrics.ConfusionMatrix;
-import org.kie.trustyai.explainability.metrics.fairness.FairnessMetricsUtils;
 import org.kie.trustyai.explainability.model.*;
+import org.kie.trustyai.metrics.accuracy.ConfusionMatrix;
+import org.kie.trustyai.metrics.fairness.FairnessMetricsUtils;
 
-public class GroupAveragePredictiveValueDifference {
+public class GroupAverageOddsDifference {
     /**
-     * Calculate average predictive value difference.
+     * Calculate average odds difference.
      *
      * @param inputSelector selector for privileged group
      * @param outputSelector selector for favorable label
      * @param dataset dataset used to evaluate AOD
      * @param model model to be evaluated fairness-wise
-     * @return average predictive value difference
+     * @return average odds difference value
      * @throws ExecutionException if any error occurs during model prediction
      * @throws InterruptedException if timeout or other interruption issues occur during model prediction
      */
@@ -30,11 +30,6 @@ public class GroupAveragePredictiveValueDifference {
         Dataset privileged = dataset.filterByInput(inputSelector);
         Map<String, Integer> privilegedCounts = FairnessMetricsUtils.countMatchingOutputSelector(privileged, model.predictAsync(privileged.getInputs()).get(), outputSelector);
 
-        double ptp = privilegedCounts.get("tp");
-        double ptn = privilegedCounts.get("tn");
-        double pfp = privilegedCounts.get("fp");
-        double pfn = privilegedCounts.get("fn");
-
         Dataset unprivileged = dataset.filterByInput(inputSelector.negate());
         Map<String, Integer> unprivilegedCounts = FairnessMetricsUtils.countMatchingOutputSelector(unprivileged, model.predictAsync(unprivileged.getInputs()).get(), outputSelector);
 
@@ -43,7 +38,12 @@ public class GroupAveragePredictiveValueDifference {
         double ufp = unprivilegedCounts.get("fp");
         double ufn = unprivilegedCounts.get("fn");
 
-        return (utp / (utp + ufp) - ptp / (ptp + pfp + 1e-10)) / 2d + (ufn / (ufn + utn) - pfn / (pfn + ptn + 1e-10)) / 2;
+        double ptp = privilegedCounts.get("tp");
+        double ptn = privilegedCounts.get("tn");
+        double pfp = privilegedCounts.get("fp");
+        double pfn = privilegedCounts.get("fn");
+
+        return (utp / (utp + ufn) - ptp / (ptp + pfn + 1e-10)) / 2d + (ufp / (ufp + utn) - pfp / (pfp + ptn + 1e-10)) / 2;
     }
 
     public static double calculate(final Dataframe samples,
@@ -60,10 +60,6 @@ public class GroupAveragePredictiveValueDifference {
         return calculate(samples, truth, priviledgeColumns, priviledgeValues, positiveClass);
     }
 
-    /**
-     * Calculate average predictive value difference.
-     *
-     */
     public static double calculate(final Dataframe test,
             final Dataframe truth,
             final List<Integer> priviledgeColumns,
@@ -96,6 +92,6 @@ public class GroupAveragePredictiveValueDifference {
         double pfp = pcm.getFalsePositives();
         double pfn = ucm.getFalseNegatives();
 
-        return (utp / (utp + ufp) - ptp / (ptp + pfp + 1e-10)) / 2d + (ufn / (ufn + utn) - pfn / (pfn + ptn + 1e-10)) / 2;
+        return (utp / (utp + ufn) - ptp / (ptp + pfn + 1e-10)) / 2d + (ufp / (ufp + utn) - pfp / (pfp + ptn + 1e-10)) / 2;
     }
 }
