+++ explainability-core/src/main/java/org/kie/trustyai/metrics/fairness/group/GroupAveragePredictiveValueDifference.java
@@ -1,4 +1,4 @@
-package org.kie.trustyai.explainability.metrics.fairness.group;
+package org.kie.trustyai.metrics.fairness.group;
 
 import java.util.List;
 import java.util.Map;
@@ -6,19 +6,19 @@ import java.util.concurrent.ExecutionException;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-import org.kie.trustyai.explainability.metrics.ConfusionMatrix;
-import org.kie.trustyai.explainability.metrics.fairness.FairnessMetricsUtils;
 import org.kie.trustyai.explainability.model.*;
+import org.kie.trustyai.metrics.accuracy.ConfusionMatrix;
+import org.kie.trustyai.metrics.fairness.FairnessMetricsUtils;
 
-public class GroupAverageOddsDifference {
+public class GroupAveragePredictiveValueDifference {
     /**
-     * Calculate average odds difference.
+     * Calculate average predictive value difference.
      *
      * @param inputSelector selector for privileged group
      * @param outputSelector selector for favorable label
      * @param dataset dataset used to evaluate AOD
      * @param model model to be evaluated fairness-wise
-     * @return average odds difference value
+     * @return average predictive value difference
      * @throws ExecutionException if any error occurs during model prediction
      * @throws InterruptedException if timeout or other interruption issues occur during model prediction
      */
@@ -30,6 +30,11 @@ public class GroupAverageOddsDifference {
         Dataset privileged = dataset.filterByInput(inputSelector);
         Map<String, Integer> privilegedCounts = FairnessMetricsUtils.countMatchingOutputSelector(privileged, model.predictAsync(privileged.getInputs()).get(), outputSelector);
 
+        double ptp = privilegedCounts.get("tp");
+        double ptn = privilegedCounts.get("tn");
+        double pfp = privilegedCounts.get("fp");
+        double pfn = privilegedCounts.get("fn");
+
         Dataset unprivileged = dataset.filterByInput(inputSelector.negate());
         Map<String, Integer> unprivilegedCounts = FairnessMetricsUtils.countMatchingOutputSelector(unprivileged, model.predictAsync(unprivileged.getInputs()).get(), outputSelector);
 
@@ -38,12 +43,7 @@ public class GroupAverageOddsDifference {
         double ufp = unprivilegedCounts.get("fp");
         double ufn = unprivilegedCounts.get("fn");
 
-        double ptp = privilegedCounts.get("tp");
-        double ptn = privilegedCounts.get("tn");
-        double pfp = privilegedCounts.get("fp");
-        double pfn = privilegedCounts.get("fn");
-
-        return (utp / (utp + ufn) - ptp / (ptp + pfn + 1e-10)) / 2d + (ufp / (ufp + utn) - pfp / (pfp + ptn + 1e-10)) / 2;
+        return (utp / (utp + ufp) - ptp / (ptp + pfp + 1e-10)) / 2d + (ufn / (ufn + utn) - pfn / (pfn + ptn + 1e-10)) / 2;
     }
 
     public static double calculate(final Dataframe samples,
@@ -60,6 +60,10 @@ public class GroupAverageOddsDifference {
         return calculate(samples, truth, priviledgeColumns, priviledgeValues, positiveClass);
     }
 
+    /**
+     * Calculate average predictive value difference.
+     *
+     */
     public static double calculate(final Dataframe test,
             final Dataframe truth,
             final List<Integer> priviledgeColumns,
@@ -92,6 +96,6 @@ public class GroupAverageOddsDifference {
         double pfp = pcm.getFalsePositives();
         double pfn = ucm.getFalseNegatives();
 
-        return (utp / (utp + ufn) - ptp / (ptp + pfn + 1e-10)) / 2d + (ufp / (ufp + utn) - pfp / (pfp + ptn + 1e-10)) / 2;
+        return (utp / (utp + ufp) - ptp / (ptp + pfp + 1e-10)) / 2d + (ufn / (ufn + utn) - pfn / (pfn + ptn + 1e-10)) / 2;
     }
 }
