+++ explainability-core/src/test/java/org/kie/trustyai/explainability/metrics/FairnessMetricsUtilsTest.java
@@ -26,15 +26,15 @@ import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.kie.trustyai.explainability.metrics.fairness.group.DisparateImpactRatio;
-import org.kie.trustyai.explainability.metrics.fairness.group.GroupAverageOddsDifference;
-import org.kie.trustyai.explainability.metrics.fairness.group.GroupAveragePredictiveValueDifference;
-import org.kie.trustyai.explainability.metrics.fairness.group.GroupStatisticalParityDifference;
-import org.kie.trustyai.explainability.metrics.fairness.individual.IndividualConsistency;
-import org.kie.trustyai.explainability.metrics.utils.FairnessDefinitions;
 import org.kie.trustyai.explainability.model.*;
 import org.kie.trustyai.explainability.utils.DataUtils;
 import org.kie.trustyai.explainability.utils.models.TestModels;
+import org.kie.trustyai.metrics.fairness.FairnessDefinitions;
+import org.kie.trustyai.metrics.fairness.group.DisparateImpactRatio;
+import org.kie.trustyai.metrics.fairness.group.GroupAverageOddsDifference;
+import org.kie.trustyai.metrics.fairness.group.GroupAveragePredictiveValueDifference;
+import org.kie.trustyai.metrics.fairness.group.GroupStatisticalParityDifference;
+import org.kie.trustyai.metrics.fairness.individual.IndividualConsistency;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
